fix of bug causing magnetic_transient_planar_profile_conductor problem not to stop. Now problem in postprocessor

+++ src/hermes2d/problem.h
@@ -50,7 +50,7 @@ public:
     inline Value timeTotal() const { return m_timeTotal; }
     void setTimeTotal(const Value &timeTotal) { m_timeTotal = timeTotal; emit changed(); }
 
-    inline int constantTimeStep() { return m_timeTotal.number() / m_numConstantTimeSteps; }
+    inline double constantTimeStep() { return m_timeTotal.number() / m_numConstantTimeSteps; }
 
     inline TimeStepMethod timeStepMethod() const {return m_timeStepMethod; }
     void setTimeStepMethod(TimeStepMethod timeStepMethod) { m_timeStepMethod = timeStepMethod; }
